refactor(discord): split handlers into _shared + per-game providers (#268)
handlers.py is now a thin dispatcher: each handle_* keeps its name/route, loads
the event, resolves get_signup_handler(event.game_type), and delegates. Rank-flow
entries guard isinstance(handler, DotaHandler). _shared.py holds _get_org_user/
_load_event/_direct_signup/_existing_active_signup/_log_*. providers/ holds
DotaHandler/DeadlockHandler/DefaultHandler + registry. game_type resolves from
event (client value ignored). events.services imports stay function-local (cycle).
Cache invalidation preserved (deadlock profile, dota set_position, event_repeater).
Fixed PositionConfirm test patch target (provider namespace, not _shared).
462 tests: only 4 pre-existing nickname failures remain (unrelated).
